requirements.txt
# Azure OpenAI
openai>=1.0.0
azure-identity>=1.12.0

# LangChain et ses dépendances
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.10

# Vector stores et embeddings
chromadb>=0.4.18
sentence-transformers>=2.2.2
torch>=2.2.1
transformers>=4.35.0
accelerate>=0.24.1
bitsandbytes>=0.41.1
huggingface-hub>=0.16.4,<0.18
tokenizers>=0.14,<0.15

# Document loaders
unstructured>=0.10.30
python-docx>=0.8.11
pdfminer.six>=20221105
markdown>=3.5.1
pypdf>=3.17.1

# Web framework
flask>=2.3.3
flask-admin>=1.6.1
flask-sqlalchemy>=3.1.1
flask-cors>=4.0.0
python-dotenv>=1.0.0

# Base de données
psycopg2-binary>=2.9.9

# Utilitaires
nltk>=3.8.1
langdetect>=1.0.9
PyJWT>=2.8.0
python-multipart>=0.0.6
requests>=2.31.0
beautifulsoup4>=4.12.2
pandas>=2.1.4

# OCR et traitement d'images
pytesseract>=0.3.10
opencv-python>=4.8.1.78

# Typage
typing-extensions>=4.8.0